+++ backend/src/feed/feed.controller.ts
@@ -3,14 +3,6 @@ import CreateFeedDto from './dto/create.feed.dto';
 import { FeedService } from './feed.service';
 import { Feed } from 'src/customDecorator/feed.decorator';
 import userIdDto from './dto/create.userId.dto';
-import {
-  createCipher,
-  createCipheriv,
-  createDecipheriv,
-  randomBytes,
-  scrypt,
-} from 'crypto';
-import { promisify } from 'util';
 
 @Controller('feed')
 export class FeedController {
@@ -22,27 +14,11 @@ export class FeedController {
     @Feed() createFeedDto: CreateFeedDto,
   ) {
     const id = new userIdDto(userId);
-    const feed = await this.feedService.createFeed(createFeedDto, id.userId);
+    const feedParam = await this.feedService.createFeed(
+      createFeedDto,
+      id.userId,
+    );
 
-    const iv = randomBytes(16);
-    const key = process.env.SECRET_KEY;
-
-    const cipher = createCipheriv('aes-256-cbc', key, iv);
-
-    const encryptedText = Buffer.concat([
-      cipher.update(feed.id.toString()),
-      cipher.final(),
-    ]).toString('hex');
-
-    const decipher = createDecipheriv('aes-256-cbc', key, iv);
-    const encrypt = Buffer.from(encryptedText, 'hex');
-    const decryptedText = Buffer.concat([
-      decipher.update(encrypt),
-      decipher.final(),
-    ]).toString();
-
-    console.log(encryptedText, decryptedText);
-
-    return encryptedText;
+    return feedParam;
   }
 }
